Reestructura de users y products a capas
